pkg/secure/chacha20poly1305.go
package secure

import (
	"crypto/rand"
	"fmt"
	"golang.org/x/crypto/chacha20poly1305"
)

// EncryptChaCha20Poly1305 encrypts the plaintext using ChaCha20-Poly1305.
func EncryptChaCha20Poly1305(key, plaintext []byte) ([]byte, error) {
	aead, err := chacha20poly1305.New(key)
	if err != nil {
		return nil, err
	}

	nonce := make([]byte, chacha20poly1305.NonceSize)
	if _, err := rand.Read(nonce); err != nil {
		return nil, err
	}

	ciphertext := aead.Seal(nonce, nonce, plaintext, nil)
	return ciphertext, nil
}

// DecryptChaCha20Poly1305 decrypts the ciphertext using ChaCha20-Poly1305.
func DecryptChaCha20Poly1305(key, ciphertext []byte) ([]byte, error) {
	aead, err := chacha20poly1305.New(key)
	if err != nil {
		return nil, err
	}

	nonceSize := chacha20poly1305.NonceSize
	if len(ciphertext) < nonceSize {
		return nil, fmt.Errorf("ciphertext too short")
	}

	nonce, ciphertext := ciphertext[:nonceSize], ciphertext[nonceSize:]
	plaintext, err := aead.Open(nil, nonce, ciphertext, nil)
	if err != nil {
		return nil, err
	}

	return plaintext, nil
}
